requirements.txt
Django==1.4.1
Jinja2==2.6
Pillow==1.7.7
Pygments==1.5
South==0.7.6
#distribute==0.6.19
django-appconf==0.5
django-autoslug==1.6.1
django-countries==1.3
django-filebrowser==3.5.0
django-grappelli==2.4.2
django-staticfiles==1.2.1
django-tinymce==1.5.1b2
docutils==0.9.1
easy-thumbnails==1.1
wsgiref==0.1.2
